requirements.txt
aioredis==2.0.1
aiosqlite==0.21.0
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
cattrs==25.1.1
cfgv==3.4.0
click==8.2.1
colorama==0.4.6
distlib==0.3.9
docstring-to-markdown==0.17
fastapi==0.115.12
filelock==3.18.0
greenlet==3.2.3
grimp==3.9
h11==0.16.0
identify==2.6.12
idna==3.10
import-linter==2.3
importlib_metadata==8.7.0
jedi==0.19.2
joblib==1.5.1
lsprotocol==2025.0.0
Mako==1.3.10
MarkupSafe==3.0.2
mypy==1.16.1
mypy_extensions==1.1.0
nodeenv==1.9.1
parso==0.8.4
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
PyJWT==2.10.1
pylsp-mypy==0.7.0
python-dotenv==1.1.0
python-lsp-jsonrpc==1.1.2
python-lsp-ruff==2.2.2
python-lsp-server==1.12.2
PyYAML==6.0.2
redis==6.2.0
ruff==0.12.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
ujson==5.10.0
uvicorn==0.34.3
virtualenv==20.31.2
zipp==3.23.0
